+++ gulpfile.js
@@ -81,6 +81,8 @@ var vendors = {
     'bower_components/angular-animate/angular-animate.js',
     // 'bower_components/angular-gamepad/dist/angular-gamepad.js', // Fixme
     'bower_components/angular-hotkeys/build/hotkeys.js',
+    'bower_components/angular-translate/angular-translate.js',
+    'bower_components/angular-translate-loader-static-files/angular-translate-loader-static-files.js',
     'bower_components/gsap/src/uncompressed/TweenLite.js',
     'bower_components/gsap/src/uncompressed/plugins/CSSPlugin.js',
     'bower_components/gsap/src/uncompressed/jquery.gsap.js',