+++ app_sources/index.js
@@ -4,7 +4,6 @@ var app = angular.module('app', [
   'ngLoad',
   'ngRoute',
   'ngResize',
-  // 'ngRaven',
   'cfp.hotkeys',
   'jsonFormatter',
   'templates',
@@ -47,36 +46,11 @@ app.config(['$translateProvider', function($translateProvider) {
   $translateProvider.fallbackLanguage('en_US');
 }]);
 
-// =========== Raven-js (sentry) config
-// app.value("RavenConfig", {
-//   dsn: 'http://a0838271e36b48b5883a9d2b6909e3a5@sentry.vik.io/2',
-//   config: {
-//     dataCallback: function(data) {
-//       // remove everything except filename
-//       data.culprit = data.culprit.split('/').pop();
-//       data.request.url = data.request.url.split('index.html#')[1];
-//       for (var i in data.stacktrace.frames) {
-//         data.stacktrace.frames[i].filename =
-//           data.stacktrace.frames[i].filename.split('/').pop();
-//       }
-//       return data;
-//     }
-//   }
-// });
-
 // =========== Preload services and language
 app.run(['$translate', 'settings', 'inputs', function($translate, settings, inputs) {
   // Force the settings service to be instanciated early
   // Apply language
   $translate.use(settings.$obj.language);
-}]);
-
-// =========== zspin core settings
-app.run(['zspin', function(zspin) {
-
-  // set Raven infos
-  // Raven.setReleaseContext(zspin.gui.App.manifest.version);
-  // Raven.setTagsContext({'nw.js': process.versions['node-webkit']});
 }]);
 
 // =========== config shortcuts