+++ gulpfile.js
@@ -89,8 +89,6 @@ var vendors = {
     'bower_components/json-formatter/dist/json-formatter.js',
     'bower_components/angular-toastr/dist/angular-toastr.tpls.js',
     'bower_components/jplayer/dist/jplayer/jquery.jplayer.js',
-    'bower_components/raven-js/dist/raven.js',
-    'bower_components/raven-js/plugins/angular.js',
   ],
   'styles': [
     'bower_components/skeleton/css/normalize.css',