+++ CHANGES.md
@@ -1,7 +1,7 @@
 ChangeLog
 =========
 
-0.2.0 (unreleased)
+0.2.0 (2015-09-08)
 ------------------
 
 * Replaced nw.js with atom-electron